+++ bw.workflow.convictedreportinquirybysearchcriteria/src/main/java/sa/gov/nic/bio/bw/workflow/convictedreportinquirybysearchcriteria/ConvictedReportInquiryBySearchCriteriaWorkflow.java
@@ -15,8 +15,8 @@ import sa.gov.nic.bio.bw.workflow.commons.lookups.CrimeTypesLookup;
 import java.util.Locale;
 import java.util.ResourceBundle;
 
-@AssociatedMenu(workflowId = 1011, menuId = "menu.query.convictedreportinquirybysearchcriteria",
-				menuTitle = "menu.title", menuOrder = 6, devices = Device.BIO_UTILITIES)
+@AssociatedMenu(workflowId = 1011, menuId = "menu.query.convictedReportInquiryBySearchCriteria",
+				menuTitle = "menu.title", menuOrder = 5, devices = Device.BIO_UTILITIES)
 @WithLookups({PersonTypesLookup.class, DocumentTypesLookup.class, CountriesLookup.class, CrimeTypesLookup.class})
 public class ConvictedReportInquiryBySearchCriteriaWorkflow extends SinglePageWorkflowBase
 {