+++ RELEASE-NOTES.md
@@ -20,6 +20,7 @@ VERSION SCHEMA: vYYYY.MM.#SEQ
 - ENH: Upgrade ScenicView.
 - CHN: Update the fingerprint inquiry results to include records from DIS.
 - CHN: Change the minimum application resolution from 800x600 to 1024x768.
+- NEW: Add Fingerprints-Inquiry workflow.
 
 ---